P02: Add screenshots and write README.md
Create a new directory to include all the screenshots of the
final result of this project. Add screenshots for both the desktop
and mobile view containing normal, empty-error, invalid-error
and final views for each.

Write a detailed README.md for this project mentioning the overview,
challenges faced and lessons learned. Also add all the screenshots
of this project of the final result.

Add the link of this project in the main index.html