Try fixing the "nav__overlay" to also overlay img
The desired result is that the "nav__overlay" class should
be applied on entire body when the side nav is active.

This could be achieved by using z-index property of the
overlay, which should be less than the z-index of the side
nav.

Regardless, the result is not achieved, even implementing
what is suggested above.